+++ src/containers/Tenant/Query/QueryResult/components/QueryInfoDropdown/useQueryInfoMenuItems.tsx
@@ -5,12 +5,15 @@ import type {DropdownMenuItem} from '@gravity-ui/uikit';
 import {Text} from '@gravity-ui/uikit';
 
 import {planToSvgApi} from '../../../../../../store/reducers/planToSvg';
-import type {QueryPlan, ScriptPlan, TKqpStatsQuery} from '../../../../../../types/api/query';
 import {cn} from '../../../../../../utils/cn';
-import createToast from '../../../../../../utils/createToast';
-import {prepareCommonErrorMessage} from '../../../../../../utils/errors';
-import {parseQueryError} from '../../../../../../utils/query';
 import i18n from '../../i18n';
+
+import {
+    createHandleDiagnosticsDownload,
+    createHandleDownload,
+    createHandleOpenInNewTab,
+} from './utils';
+import type {QueryResultsInfo} from './utils';
 const b = cn('query-info-dropdown');
 
 export interface MenuItemContentProps {
@@ -27,19 +30,6 @@ export function MenuItemContent({title, description}: MenuItemContentProps) {
             </Text>
         </div>
     );
-}
-
-export interface QueryResultsInfo {
-    ast?: string;
-    stats?: TKqpStatsQuery;
-    queryText?: string;
-    plan?: QueryPlan | ScriptPlan;
-}
-
-export interface DiagnosticsData extends QueryResultsInfo {
-    database: string;
-    timestamp: string;
-    error?: ReturnType<typeof parseQueryError>;
 }
 
 export interface UseQueryInfoMenuItemsProps {
@@ -71,61 +61,24 @@ export function useQueryInfoMenuItems({
 
         const plan = queryResultsInfo.plan;
         if (plan && hasPlanToSvg) {
-            const handleGetSvg = () => {
-                if (blobUrl) {
-                    return Promise.resolve(blobUrl);
-                }
-
-                return getPlanToSvg({plan, database})
-                    .unwrap()
-                    .then((result) => {
-                        const blob = new Blob([result], {type: 'image/svg+xml'});
-                        const url = URL.createObjectURL(blob);
-                        setBlobUrl(url);
-                        return url;
-                    })
-                    .catch((err) => {
-                        const errorMessage = prepareCommonErrorMessage(err);
-                        createToast({
-                            title: i18n('text_error-plan-svg', {error: errorMessage}),
-                            name: 'plan-svg-error',
-                            type: 'error',
-                        });
-                        return null;
-                    });
-            };
-
-            const handleOpenInNewTab = () => {
-                handleGetSvg().then((url) => {
-                    if (url) {
-                        window.open(url, '_blank');
-                    }
-                });
-            };
-
-            const handleDownload = () => {
-                handleGetSvg().then((url) => {
-                    const link = document.createElement('a');
-                    if (url) {
-                        link.href = url;
-                        link.download = 'query-plan.svg';
-                        document.body.appendChild(link);
-                        link.click();
-                        document.body.removeChild(link);
-                    }
-                });
+            const handlersParams = {
+                plan,
+                database,
+                blobUrl,
+                getPlanToSvg,
+                setBlobUrl,
             };
 
             menuItems.push([
                 {
                     text: (
                         <MenuItemContent
-                            title={i18n('text_open-new-tab')}
-                            description={i18n('text_open-new-tab_description')}
+                            title={i18n('text_open-execution-plan')}
+                            description={i18n('text_open-execution-plan_description')}
                         />
                     ),
                     icon: <ArrowUpRightFromSquare className={b('icon')} />,
-                    action: handleOpenInNewTab,
+                    action: createHandleOpenInNewTab(handlersParams),
                     className: b('menu-item'),
                 },
                 {
@@ -136,35 +89,13 @@ export function useQueryInfoMenuItems({
                         />
                     ),
                     icon: <ArrowDownToLine className={b('icon')} />,
-                    action: handleDownload,
+                    action: createHandleDownload(handlersParams),
                     className: b('menu-item'),
                 },
             ]);
         }
 
         if (queryResultsInfo) {
-            const handleDiagnosticsDownload = () => {
-                const parsedError = error ? parseQueryError(error) : undefined;
-                const diagnosticsData: DiagnosticsData = {
-                    ...queryResultsInfo,
-                    database,
-                    timestamp: new Date().toISOString(),
-                    ...(parsedError && {error: parsedError}),
-                };
-
-                const blob = new Blob([JSON.stringify(diagnosticsData, null, 2)], {
-                    type: 'application/json',
-                });
-                const url = URL.createObjectURL(blob);
-                const link = document.createElement('a');
-                link.href = url;
-                link.download = `query-diagnostics-${new Date().getTime()}.json`;
-                document.body.appendChild(link);
-                link.click();
-                document.body.removeChild(link);
-                URL.revokeObjectURL(url);
-            };
-
             menuItems.push([
                 {
                     text: (
@@ -174,14 +105,14 @@ export function useQueryInfoMenuItems({
                         />
                     ),
                     icon: <ArrowDownToLine className={b('icon')} />,
-                    action: handleDiagnosticsDownload,
+                    action: createHandleDiagnosticsDownload({queryResultsInfo, database, error}),
                     className: b('menu-item'),
                 },
             ]);
         }
 
         return menuItems;
-    }, [queryResultsInfo, hasPlanToSvg, blobUrl, getPlanToSvg, database]);
+    }, [queryResultsInfo, hasPlanToSvg, database, blobUrl, getPlanToSvg, error]);
 
     return {
         isLoading,
