+++ src/containers/Tenant/Query/QueryResult/i18n/en.json
@@ -11,10 +11,9 @@
   "trace": "Trace",
   "title.truncated": "Truncated",
   "title.result": "Result",
-  "text_plan-svg": "Execution plan",
-  "text_open-new-tab": "Open in new tab",
-  "text_open-new-tab_description": "New tab",
-  "text_download": "Download",
+  "text_open-execution-plan": "Open Execution Plan",
+  "text_open-execution-plan_description": "New tab",
+  "text_download": "Download Execution Plan",
   "text_download_description": "SVG",
   "text_diagnostics": "Download Diagnostics",
   "text_diagnostics_description": "JSON",
